app/requirements.txt
appnope==0.1.4
astroid==3.1.0
asttokens==2.4.1
autopep8==2.1.0
black==24.4.0
blinker==1.7.0
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
comm==0.2.2
crashtest==0.4.1
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
Flask==2.1.3
Flask-Login==0.6.3
Flask-Script==2.0.5
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.21.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
idna==3.7
installer==0.7.0
ipykernel==6.29.4
ipython==8.24.0
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==24.3.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.1
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.19.1
Pygments==2.17.2
pylint==3.1.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
pyzmq==26.0.3
rapidfuzz==3.8.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
stack-data==0.6.3
tomlkit==0.12.4
tornado==6.4
traitlets==5.14.3
trove-classifiers==2024.4.10
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==2.2.2
wheel==0.43.0
WTForms==3.1.2
wtforms-html5==0.6.1
xattr==1.1.0
